// src/domain/enums/MedalType.ts
export enum MedalType {
  Accuracy = 'Accuracy',
  Assists = 'Assists',
  Captures = 'Captures',
  ComboKill = 'Combo Kill',
  Defends = 'Defends',
  Excellent = 'Excellent',
  FirstFrag = 'First Frag',
  Headshot = 'Headshot',
  Humiliation = 'Humiliation',
  Impressive = 'Impressive',
  Midair = 'Midair',
  Perfect = 'Perfect',
  Perforated = 'Perforated',
  QuadGod = 'Quad God',
  Rampage = 'Rampage',
  Revenge = 'Revenge'
}